requirements.txt
langchain>=0.3.0

langchain-openai>=0.2.0

langchain-community>=0.3.0

langgraph>=0.2.0

faiss-cpu>=1.8.0

python-dotenv>=1.0.0

pydantic>=2.0.0
